!Keine Änderungen! Nur Code neu formatiert

+++ core/src/de/ft/robocontrol/utils/CheckKollision.java
@@ -9,7 +9,7 @@ import de.ft.robocontrol.MainGame;
 
 public class CheckKollision {
 
-    private static int mousesize=0;
+    private static int mousesize = 0;
 
     public static boolean object(int obj1_x, int obj1_y, int obj1_h, int obj1_w, int obj2_x, int obj2_y, int obj2_h, int obj2_w) {
 
@@ -23,22 +23,22 @@ public class CheckKollision {
     public static boolean checkmousewithblock(Block block) {
 
 
-            return object((int) block.getX(), (int) block.getY(), (int) block.getH(), (int) block.getW(), (int)MainGame.viewport.unproject(new Vector3(Gdx.input.getX(), Gdx.input.getY(),0)).x, (int) MainGame.viewport.unproject(new Vector3(Gdx.input.getX(), Gdx.input.getY(),0)).y, (int) mousesize, (int) mousesize);
+        return object(block.getX(), block.getY(), block.getH(), block.getW(), (int) MainGame.viewport.unproject(new Vector3(Gdx.input.getX(), Gdx.input.getY(), 0)).x, (int) MainGame.viewport.unproject(new Vector3(Gdx.input.getX(), Gdx.input.getY(), 0)).y, mousesize, mousesize);
 
-        }
+    }
 
 
     public static boolean checkmousewithobject(int obj1_x, int obj1_y, int obj1_h, int obj1_w, Vector2 mousepos) {
 
 
-        return object(obj1_x, obj1_y, obj1_w, obj1_h, (int) mousepos.x, (int)mousepos.y,  (int) mousesize, (int) mousesize);
+        return object(obj1_x, obj1_y, obj1_w, obj1_h, (int) mousepos.x, (int) mousepos.y, mousesize, mousesize);
 
     }
 
     public static boolean checkmousewithobject(int obj1_x, int obj1_y, int obj1_h, int obj1_w, int x, int y) {
 
 
-        return object(obj1_x, obj1_y, obj1_w, obj1_h, (int) x, (int)y,  (int) mousesize, (int) mousesize);
+        return object(obj1_x, obj1_y, obj1_w, obj1_h, x, y, mousesize, mousesize);
 
     }
 
@@ -46,36 +46,36 @@ public class CheckKollision {
     public static boolean checkmousewithblock(Block block, Vector2 mousepos) {
 
 
-        return object((int) block.getX(), (int) block.getY(), (int) block.getH(), (int) block.getW(), (int) mousepos.x, (int)mousepos.y,  (int) mousesize, (int) mousesize);
+        return object(block.getX(), block.getY(), block.getH(), block.getW(), (int) mousepos.x, (int) mousepos.y, mousesize, mousesize);
 
     }
 
-    public static boolean checkblockwithduplicate(Block normal, Block duplicate,int rl) {
-        if(rl==0) {
+    public static boolean checkblockwithduplicate(Block normal, Block duplicate, int rl) {
+        if (rl == 0) {
             return object(normal.getX(), normal.getY(), normal.getH(), normal.getW(), duplicate.getX_dup_rechts(), duplicate.getY(), duplicate.getH(), 60);
-        }else{
-            return object(normal.getX(), normal.getY(), normal.getH(), normal.getW(), duplicate.getX_dup_links()+duplicate.getW()-60, duplicate.getY(), duplicate.getH(), 60);
+        } else {
+            return object(normal.getX(), normal.getY(), normal.getH(), normal.getW(), duplicate.getX_dup_links() + duplicate.getW() - 60, duplicate.getY(), duplicate.getH(), 60);
         }
     }
 
 
-    public static int flache(int x,int y,int w,int h,int xm,int ym){
-        int flache=0;
+    public static int flache(int x, int y, int w, int h, int xm, int ym) {
+        int flache = 0;
 
         int xx = w - (xm - x);
         int yy = (ym + h - y);
-        if(xx>w){
-            xx=w-(xx-w);
+        if (xx > w) {
+            xx = w - (xx - w);
         }
-        if(yy>h){
-            yy=h-(yy-h);
+        if (yy > h) {
+            yy = h - (yy - h);
         }
 
         //System.out.println("y: " + yy + "x: " + xx + "flache: "+ yy*xx);
-        flache=xx*yy;
+        flache = xx * yy;
         return flache;
     }
 
 
-    }
+}
